requirements.txt
git+https://github.com/Rapptz/discord.py
async-timeout==3.0.1
discord.py
requests==2.22.0
python-dotenv